backend/ai/text_generation.py
import transformers

REQUESTS_CA_BUNDLE=r'/etc/ssl/certs/ca-certificates.crt'

class TextGeneration():
    def __init__(self):
        pass


    def generate_text(self, prompt):
        
        input_ids = None

        # Generate text using the Llama Transformer model
        outputs = self.model.generate(
            input_ids=input_ids,
            max_length=self.max_length,
            temperature=self.temperature,
            num_beams=self.num_beams,
            do_sample=True,
            top_k=50,
            top_p=0.95,
            repetition_penalty=1.0,
            length_penalty=1.0,
            pad_token_id=self.tokenizer.pad_token_id,
            eos_token_id=self.tokenizer.eos_token_id,
            bos_token_id=self.tokenizer.bos_token_id,
            num_return_sequences=1
        )

        # Decode the generated text using the BERT tokenizer
        generated_text = self.tokenizer.decode(outputs[0], skip_special_tokens=True)

        # Print the generated text
        print(generated_text)
        return generated_text