requirements.txt
black==23.3.0
click>=8.0.0
mypy==1.3.0
openai==0.27.8
pre-commit==3.3.3
pytest==7.3.1
ruff==0.0.272
termcolor==2.3.0
typer>=0.3.2
rudder-sdk-python==2.0.2
dataclasses-json==0.5.7
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
Flask==2.3.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
bcrypt==4.0.1
flask_wtf>=1.1.1
email_validator>=2.0.0
googlemaps==4.10.0
pillow==10.0.0
joblib==1.3.2
pandas==2.1.0
numpy>=1.23.2
django==4.2.5
scikit-learn==1.3.1
fast_autocomplete==0.9.0
python-Levenshtein>=0.12.2
wtforms_sqlalchemy==0.4.1
Werkzeug==2.3.0
